Created ArraySplitter Class. Creted Unit Tests for Array Splitter
The array splitter class will take in a provided array and split it into equal sized groups. The number of equal sized groups is determined by a number of splits provided to the method. If there would be any remainders that would not fit into an equal sized group, they are put in their own group at the end of the array. Unittest test cases have also been produced to test the functionality of this class.